feat: completar ciclo del partido, implementar DTOs y corregir recursividad
- BACKEND: Implementado 'AlineacionResponseDto' para evitar bucle infinito de Hibernate.
- BACKEND: Actualizado 'AlineacionController' para devolver datos planos (sin anidamiento).
- BACKEND: Corregidos tipos (Long/Integer) en 'JugadorController' y DTOs.
- FRONTEND: Actualizado 'match-detail' para consumir el nuevo DTO plano (nombre, foto, goles).
- FRONTEND: Reparada la página 'edit-match' (Entrenador) para guardar acta correctamente.
- FRONTEND: Dashboard de Jugador ahora muestra estadísticas (goles/minutos) reales calculadas desde SQL.
- UI: Mejoras visuales en listas de convocatoria y manejo de imágenes rotas (placeholders).

+++ src/backend-tfg/backend-tfg/src/main/java/com/DAMUnitedFC/backend_tfg/dto/JugadorDto.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class JugadorDto {
-    private Integer idUsuario;
+    private Long idUsuario; // 🔥 Corregido a Long para coincidir con Usuario
     private String fechaNacimiento;
     private String posicion;
     private Integer dorsal;
@@ -14,5 +14,5 @@ public class JugadorDto {
     private String fechaAlta;
     private String fechaBaja;
     private String observaciones;
-    private Integer equipoPrincipal; // FK
+    private Long equipoPrincipal; // 🔥 Corregido a Long para coincidir con Equipo
 }
